// angular/src/services/interfaces/models.ts
export interface AgingBalanceReportDTO {
    customerId?: number;
    accountNumber?: number;
    customerName?: string | undefined;
    paymentTermId?: number;
    paymentTermName?: string | undefined;
    state?: string | undefined;
    phoneNumber?: string | undefined;
    email?: string | undefined;
    current30Days?: number | undefined;
    over30Days?: number | undefined;
    over60Days?: number | undefined;
    over90Days?: number | undefined;
    totalOwed?: number | undefined;
    lastPaymentDate?: moment.Moment | undefined;
    lastUpdateDate?: moment.Moment | undefined;
}
// ...
export interface Contact {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    locationId?: number | undefined;
    location?: Location;
    positionTypeId?: number | undefined;
    positionType?: PositionType;
    contactName?: string | undefined;
    phoneNumber?: string | undefined;
    email?: string | undefined;
    notes?: string | undefined;
    isEmailOrder?: boolean;
    isEmailCreditMemo?: boolean;
    isEmailStatement?: boolean;
}

export interface CreateOrderResult {
    order?: Order;
    orderNumber?: number | undefined;
    quoteNumber?: number | undefined;
    orderResult?: boolean;
    syncResult?: boolean | undefined;
    isQuote?: boolean;
}

export interface Customer {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    latitude?: string | undefined;
    longitude?: string | undefined;
    accountNumber?: number;
    customerName?: string | undefined;
    customerTypeId?: number;
    customerType?: CustomerType;
    priceLevelId?: number | undefined;
    priceLevel?: PriceLevel;
    paymentTermId?: number | undefined;
    paymentTerm?: PaymentTerm;
    salesRepresentativeOutId?: number | undefined;
    salesRepresentativeOut?: SalesRepresentative;
    salesRepresentativeInId?: number | undefined;
    salesRepresentativeIn?: SalesRepresentative;
    defaultWarehouseId?: number | undefined;
    warehouse?: Warehouse;
    contactName?: string | undefined;
    taxRate?: number | undefined;
    overBalance?: number | undefined;
    creditLimit?: number | undefined;
    isBypassCreditLimit?: boolean;
    isTaxable?: boolean;
    isHoldAccount?: boolean;
    discount?: number | undefined;
    sellersPermit?: string | undefined;
    trialBalance?: number | undefined;
    accountsReceivableBalance?: number | undefined;
    crossStreet?: string | undefined;
    isWithShippingCost?: boolean;
    isWithHandlingCost?: boolean;
    isAllowViewCA?: boolean;
    isAllowViewVA?: boolean;
    isAllowViewWAG?: boolean;
    isOverrideCreditLimitRestriction?: boolean;
    shipperAccountNumber?: string | undefined;
    isApplyPFGShippingCharge?: boolean;
    isApplyPFGHandlingCharge?: boolean;
    isPFGShippingHandlingLevel?: boolean;
    freightPrepaidAmount?: number | undefined;
}

export interface CustomerCredit {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    customerId?: number;
    customer?: Customer;
    referenceId?: number | undefined;
    creditType?: number;
    creditReason?: string | undefined;
    previousRecordId?: number;
    amountPostedDate?: moment.Moment;
    currentBalance?: number;
    previousBalance?: number;
    amountPosted?: number;
}

export interface CustomerDTO {
    id?: number;
    accountNumber?: number;
    customerName?: string | undefined;
    discount?: number | undefined;
    taxRate?: number | undefined;
    paymentTermId?: number | undefined;
    priceLevelId?: number | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    zone?: string | undefined;
    contactName?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    isHoldAccount?: boolean;
    creditLimit?: number;
    isBypassCreditLimit?: boolean;
    overBalance?: number;
    statementAmount?: number;
    contacts?: Contact[] | undefined;
    customerCredit?: CustomerCredit;
    locations?: Location[] | undefined;
}
// ...
export interface CustomerSalesAmountDTO {
    customerId?: number;
    accountNumber?: number;
    customerName?: string | undefined;
    phoneNumber?: string | undefined;
    state?: string | undefined;
    column1?: SalesData;
    column2?: SalesData;
    column3?: SalesData;
    column4?: SalesData;
    column5?: SalesData;
    column6?: SalesData;
    column7?: SalesData;
    column8?: SalesData;
    column9?: SalesData;
    column10?: SalesData;
    column11?: SalesData;
    column12?: SalesData;
}

export interface CustomerSalesFilterDTO {
    col1FrDate?: moment.Moment | undefined;
    col1ToDate?: moment.Moment | undefined;
    col2FrDate?: moment.Moment | undefined;
    col2ToDate?: moment.Moment | undefined;
    col3FrDate?: moment.Moment | undefined;
    col3ToDate?: moment.Moment | undefined;
    col4FrDate?: moment.Moment | undefined;
    col4ToDate?: moment.Moment | undefined;
    col5FrDate?: moment.Moment | undefined;
    col5ToDate?: moment.Moment | undefined;
    col6FrDate?: moment.Moment | undefined;
    col6ToDate?: moment.Moment | undefined;
    col7FrDate?: moment.Moment | undefined;
    col7ToDate?: moment.Moment | undefined;
    col8FrDate?: moment.Moment | undefined;
    col8ToDate?: moment.Moment | undefined;
    col9FrDate?: moment.Moment | undefined;
    col9ToDate?: moment.Moment | undefined;
    col10FrDate?: moment.Moment | undefined;
    col10ToDate?: moment.Moment | undefined;
    col11FrDate?: moment.Moment | undefined;
    col11ToDate?: moment.Moment | undefined;
    col12FrDate?: moment.Moment | undefined;
    col12ToDate?: moment.Moment | undefined;
}

export interface CustomerType {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
}
// ...
export interface Location {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    faxNumber?: string | undefined;
    email?: string | undefined;
    latitude?: string | undefined;
    longitude?: string | undefined;
    customerId?: number;
    customer?: Customer;
    locationTypeId?: number;
    locationType?: LocationType;
    locationName?: string | undefined;
    locationCode?: string | undefined;
    notes?: string | undefined;
    zone?: string | undefined;
}
// ...
export interface LocationType {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
}
// ...
export interface PaymentTerm {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    termName?: string | undefined;
    description?: string | undefined;
    numberOfDays?: number;
    notes?: string | undefined;
}
// ...
export interface PositionType {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
}
// ...
export interface Role {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
    userTypeId?: number;
    userType?: UserType;
    notes?: string | undefined;
    rolePermissions?: RolePermission[] | undefined;
}
// ...
export interface RolePermission {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    roleId?: number;
    role?: Role;
    moduleGroupId?: number;
    moduleGroup?: ModuleGroup;
    moduleId?: number;
    module?: Module;
    accessTypeId?: number;
    accessType?: AccessType;
}

export interface SalesData {
    amount?: number | undefined;
    cost?: number | undefined;
    profit?: number | undefined;
    margin?: number | undefined;
}
// ...
export interface StatementReportDTO {
    customerId?: number;
    accountNumber?: number;
    totalDue?: number;
    customerName?: string | undefined;
    paymentTermId?: number;
    paymentTermName?: string | undefined;
    addressLine1?: string | undefined;
    addressLine2?: string | undefined;
    city?: string | undefined;
    state?: string | undefined;
    country?: string | undefined;
    zipCode?: string | undefined;
    phoneNumber?: string | undefined;
    orders?: StatementReportOrderDTO[] | undefined;
}

export interface StatementReportOrderDTO {
    orderId?: number;
    orderNumber?: number | undefined;
    invoiceNumber?: number | undefined;
    purchaseOrderNumber?: string | undefined;
    totalAmount?: number | undefined;
    balance?: number | undefined;
    orderDate?: moment.Moment;
    orderType?: string | undefined;
}
// ...
export interface User {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    roleId?: number;
    role?: Role;
    customerId?: number | undefined;
    customer?: Customer;
    locationId?: number | undefined;
    location?: Location;
    userName?: string | undefined;
    email?: string | undefined;
    isCustomerUser?: boolean;
    isActivated?: boolean;
    isChangePasswordOnLogin?: boolean;
    passwordHash?: string | undefined;
    passwordSalt?: string | undefined;
}
// ...
export interface UserType {
    id?: number;
    isActive?: boolean;
    isDeleted?: boolean;
    createdBy?: string | undefined;
    createdDate?: moment.Moment;
    modifiedBy?: string | undefined;
    modifiedDate?: moment.Moment | undefined;
    name?: string | undefined;
    code?: string | undefined;
    description?: string | undefined;
    sortOrder?: number;
}